# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample api server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api",
    "paths": {
...
        "/accounts/{userId}": {
            "get": {
                "description": "get account by user id func get account infomation by userid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "get account by user id func get account infomation by userid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ResponseSuccessAccount"
                            }
                        }
                    }
                }
            }
        },
...
        "/users/delete/{id}": {
            "delete": {
                "description": "delete user by user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Delete user by user id",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ResponseError"
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/users/update": {
            "patch": {
                "description": "update user detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "update user detail",
                "parameters": [
                    {
                        "description": "User Update Data",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ResponseSuccessUser"
                            }
                        }
                    }
                }
            }
        },
        "/users/update/{id}": {
            "patch": {
                "description": "update user detail by user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "update user detail by user id",
                "parameters": [
                    {
                        "description": "User Update Data",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ResponseSuccessUser"
                            }
                        }
                    }
                }
            }
        },
        "/users/updatepassword/{id}": {
            "patch": {
                "description": "update user password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "update user password",
                "parameters": [
                    {
                        "description": "Update user password",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserPasswordUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ResponseSuccessUser"
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "get user infomation by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "get user infomation by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ResponseSuccessUser"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Account": {
            "type": "object",
            "properties": {
                "accountToken": {
                    "type": "string"
                },
                "accountType": {
                    "type": "string"
                },
                "active": {
                    "type": "boolean"
                },
                "balance": {
                    "type": "integer"
                },
                "blockFlag": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastEntryLocId": {
                    "type": "integer"
                },
                "lastEntrySpId": {
                    "type": "integer"
                },
                "lastEntryTime": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "txnHistories": {
                    "description": "One-to-many relationship with TxnHistory",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TxnHistory"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "description": "Association",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "userID": {
                    "description": "Foreign key, unique constraint for one-to-one relationship",
                    "type": "string"
                }
            }
        },
        "models.AccountResponse": {
            "type": "object",
            "properties": {
                "account_token": {
                    "type": "string"
                },
                "account_type": {
                    "type": "string"
                },
                "active": {
                    "type": "boolean"
                },
                "balance": {
                    "type": "integer"
                },
                "block_flag": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_entry_loc_id": {
                    "type": "integer"
                },
                "last_entry_sp_id": {
                    "type": "integer"
                },
                "last_entry_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "txn_histories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TxnHistory"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "UserID         uuid.UUID    `json:\"user_id\"`",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.UserResponse"
                        }
                    ]
                }
            }
        },
...
        "models.ResponseError": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string"
                        }
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.ResponseSuccessAccount": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "account": {
                            "$ref": "#/definitions/models.AccountResponse"
                        },
                        "txn_histories": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TxnHistory"
                            }
                        },
                        "user": {
                            "$ref": "#/definitions/models.UserResponse"
                        }
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.ResponseSuccessUser": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "user": {
                            "$ref": "#/definitions/models.UserResponse"
                        }
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.TxnHistory": {
            "type": "object",
            "required": [
                "equipmentNumber",
                "locEntryId",
                "locExitId",
                "spId",
                "txnAmount",
                "txnRefId",
                "txnTypeId"
            ],
            "properties": {
                "account": {
                    "description": "Association",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Account"
                        }
                    ]
                },
                "accountID": {
                    "description": "Foreign key to Account",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "equipmentNumber": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "locEntryId": {
                    "type": "integer"
                },
                "locExitId": {
                    "type": "integer"
                },
                "spId": {
                    "type": "integer"
                },
                "txnAmount": {
                    "type": "integer"
                },
                "txnRefId": {
                    "type": "string"
                },
                "txnTypeId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "One-to-one relationship with Account",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Account"
                        }
                    ]
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
        "models.UserPasswordUpdate": {
            "type": "object",
            "properties": {
                "confirm_new_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
        "models.UserResponse": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "models.UserUpdate": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}